# docs/homework_openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      properties:
        message:
          example: 操作失败
          type: string
        success:
          example: false
          type: boolean
      type: object
...
    ExcellentHomework:
      properties:
        content:
          example: /path/to/file
          type: string
        file_url:
          example: /api/v2/files/serve/123
          type: string
        id:
          example: 1
          type: integer
        imgurl:
          example: /api/v2/files/serve/789
          type: string
        like:
          example: true
          type: boolean
        likes_count:
          example: 10
          type: integer
        score:
          example: A
          type: string
        student_name:
          example: 张三
          type: string
        time:
          example: '2024-01-01 10:00:00'
          format: date-time
          type: string
        title:
          example: 数字特效的平时作业1
          type: string
      type: object
    Homework:
      properties:
        course_name:
          example: 数字特效
          type: string
        id:
          example: 1
          type: integer
        status:
          enum:
          - draft
          - submitted
          - returned
          - excellent
          - problem
          example: submitted
          type: string
        submitted_at:
          example: '2024-01-01 10:00:00'
          format: date-time
          type: string
        type_name:
          example: 平时作业1
          type: string
        version:
          example: 1
          type: integer
      type: object
...
    SuccessResponse:
      properties:
        data:
          type: object
        message:
          example: 操作成功
          type: string
        success:
          example: true
          type: boolean
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 'JWT Token认证，需要在请求头中携带 Authorization: Bearer <token>'
      scheme: bearer
      type: http
info:
  description: 作业库系统后端接口文档，包含学生端和教师端的所有作业相关功能
  title: 作业库系统API
  version: 1.0.0
openapi: 3.0.0
paths:
  /api/v2/homework/:
    get:
      description: 学生获取自己的作业列表（仅学生）
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/Homework'
                      type: array
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取我的作业列表
      tags:
      - 学生端作业
    post:
      description: 学生提交作业（仅学生）
      requestBody:
        content:
          application/json:
            schema:
              properties:
                content:
                  description: 作业内容说明
                  example: 这是我的作业说明
                  type: string
                course_id:
                  description: 课程ID
                  example: 1
                  type: integer
                file_id:
                  description: 文件ID
                  example: 123
                  type: integer
                homework_type_id:
                  description: 作业类型ID
                  example: 1
                  type: integer
              required:
              - course_id
              - homework_type_id
              - file_id
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      properties:
                        id:
                          example: 1
                          type: integer
                      type: object
                  type: object
          description: 创建成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 创建作业
      tags:
      - 学生端作业
...
  /api/v2/homework/excellent:
    get:
      description: 获取优秀作业列表（学生和教师都可以访问）
      parameters:
      - description: 页码
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: 每页数量
        in: query
        name: per_page
        schema:
          default: 10
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      properties:
                        data:
                          items:
                            $ref: '#/components/schemas/ExcellentHomework'
                          type: array
                        page:
                          example: 1
                          type: integer
                        per_page:
                          example: 10
                          type: integer
                        total:
                          example: 100
                          type: integer
                      type: object
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取优秀作业列表
      tags:
      - 优秀作业
  /api/v2/homework/excellent/{excellent_id}/like:
    post:
      description: 点赞/取消点赞优秀作业（学生和教师都可以操作）
      parameters:
      - description: 优秀作业ID
        in: path
        name: excellent_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      properties:
                        is_liked:
                          example: true
                          type: boolean
                      type: object
                  type: object
          description: 操作成功
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 优秀作业不存在
      security:
      - bearerAuth: []
      summary: 点赞/取消点赞优秀作业
      tags:
      - 优秀作业
  /api/v2/homework/favorites:
    get:
      description: 获取用户收藏的优秀作业列表（学生和教师都可以访问）
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/ExcellentHomework'
                      type: array
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取收藏列表
      tags:
      - 收藏管理
  /api/v2/homework/student:
    get:
      description: 学生获取自己的作业列表（仅学生）
      parameters:
      - description: 课程ID（可选）
        in: query
        name: course_id
        schema:
          type: integer
      - description: 作业状态（可选）
        in: query
        name: status
        schema:
          enum:
          - draft
          - submitted
          - returned
          - excellent
          - problem
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/Homework'
                      type: array
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取学生作业列表
      tags:
      - 学生端作业
  /api/v2/homework/types:
    get:
      description: 教师获取自己发布的作业类型列表，学生获取可提交的作业类型列表
      parameters:
      - description: 课程ID（可选）
        in: query
        name: course_id
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      items:
                        $ref: '#/components/schemas/HomeworkType'
                      type: array
                  type: object
          description: 获取成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 获取作业类型列表
      tags:
      - 作业类型管理
    post:
      description: |
        教师发布作业类型（仅教师）

        **课程处理逻辑**：
        - **优先使用 course_id**：如果提供了课程ID，直接使用该课程（保持向后兼容）
        - **自动检测 course_name**：如果提供了课程名称，系统会自动查找或创建课程
          - 存在同名课程：直接使用该课程ID
          - 不存在同名课程：自动创建新课程并返回课程ID

        **参数说明**：
        - `course_id` 和 `course_name` 二选一
        - 推荐使用 `course_name`，系统自动处理课程复用
      requestBody:
        content:
          application/json:
            schema:
              properties:
                content:
                  description: 作业内容
                  example: 完成数字特效的基础练习
                  type: string
                course_id:
                  description: 课程ID（可选，与course_name二选一）
                  example: 1
                  type: integer
                course_name:
                  description: 课程名称（可选，与course_id二选一。如果不存在会自动创建）
                  example: 数字特效
                  type: string
                name:
                  description: 作业类型名称
                  example: 平时作业1
                  type: string
              required:
              - name
              - content
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/SuccessResponse'
                - properties:
                    data:
                      properties:
                        course_id:
                          description: 课程ID
                          example: 1
                          type: integer
                        course_name:
                          description: 课程名称
                          example: 数字特效
                          type: string
                        id:
                          description: 作业类型ID
                          example: 1
                          type: integer
                      type: object
                  type: object
          description: 发布成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
      security:
      - bearerAuth: []
      summary: 发布作业类型
      tags:
      - 作业类型管理
  /api/v2/homework/types/{type_id}:
    delete:
      description: 删除/下架作业类型（仅教师）
      parameters:
      - description: 作业类型ID
        in: path
        name: type_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 删除成功
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 作业类型不存在
      security:
      - bearerAuth: []
      summary: 删除/下架作业类型
      tags:
      - 作业类型管理
    put:
      description: 修改作业类型信息（仅教师）
      parameters:
      - description: 作业类型ID
        in: path
        name: type_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                content:
                  example: 完成数字特效的基础练习
                  type: string
                is_active:
                  example: true
                  type: boolean
                name:
                  example: 平时作业1
                  type: string
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 修改成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 作业类型不存在
      security:
      - bearerAuth: []
      summary: 修改作业类型
      tags:
      - 作业类型管理
  /api/v2/homework/{homework_id}:
    put:
      description: 更新作业（仅学生）
      parameters:
      - description: 作业ID
        in: path
        name: homework_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                file_id:
                  description: 文件ID
                  example: 456
                  type: integer
                status:
                  description: 作业状态
                  example: submitted
                  type: string
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 更新成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 作业不存在
      security:
      - bearerAuth: []
      summary: 更新作业
      tags:
      - 学生端作业
  /api/v2/homework/{homework_id}/evaluate:
    post:
      description: 教师批改作业（仅教师）
      parameters:
      - description: 作业ID
        in: path
        name: homework_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                evaluation:
                  description: 教师评语
                  example: 优秀的作品
                  type: string
                is_excellent:
                  description: 是否标记为优秀作业
                  example: false
                  type: boolean
                is_problem:
                  description: 是否标记为问题作业
                  example: false
                  type: boolean
                is_redo:
                  description: 是否要求重做
                  example: false
                  type: boolean
                score:
                  description: 分数
                  example: A
                  type: string
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 批改成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 作业不存在
      security:
      - bearerAuth: []
      summary: 批改作业
      tags:
      - 教师端作业
  /api/v2/homework/{homework_id}/submit:
    post:
      description: 学生提交作业（仅学生）
      parameters:
      - description: 作业ID
        in: path
        name: homework_id
        required: true
        schema:
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: 提交成功
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 请求参数错误
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 权限不足
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: 作业不存在
      security:
      - bearerAuth: []
      summary: 提交作业
      tags:
      - 学生端作业
...
servers:
- description: Local development server
  url: http://localhost:5000
- description: Local network server
  url: http://192.168.8.190:5000
...